Refactor and create exhaustive polygon contains tests.
