+++ python/cuspatial/cuspatial/tests/test_contains.py
@@ -1,45 +1,118 @@
 import geopandas as gpd
-import pandas as pd
+import pytest
 from shapely.geometry import Point, Polygon
 
 import cuspatial
 
-
-def test_point_shared_with_polygon():
-    point_series = cuspatial.from_geopandas(
-        gpd.GeoSeries(
-            [
-                Point([0.25, 0.5]),
-                Point([1, 1]),
-                Point([0.5, 0.25]),
-            ]
-        )
-    )
-    polygon_series = cuspatial.from_geopandas(
-        gpd.GeoSeries(
-            [
-                Polygon([[0, 0], [0, 1], [1, 1], [0, 0]]),
-                Polygon([[0, 0], [1, 0], [1, 1], [0, 0]]),
-                Polygon([[0, 0], [1, 0], [1, 1], [0, 1], [0, 0]]),
-            ]
-        )
-    )
-    gpdpoint = point_series.to_pandas()
-    gpdpolygon = polygon_series.to_pandas()
-    pd.testing.assert_series_equal(
-        gpdpolygon.contains(gpdpoint),
-        polygon_series.contains(point_series).to_pandas(),
-    )
+clockwiseTriangle = Polygon([[0, 0], [0, 1], [1, 1], [0, 0]])
+clockwiseSquare = Polygon(
+    [[-0.5, -0.5], [-0.5, 0.5], [0.5, 0.5], [0.5, -0.5], [-0.5, -0.5]]
+)
 
 
-def test_point_collinear_with_polygon():
-    point = Point([0.5, 0.0])
-    polygon = Polygon([[0, 0], [0, 1], [1, 1], [0, 0]])
-    point_series = cuspatial.from_geopandas(gpd.GeoSeries(point))
-    polygon_series = cuspatial.from_geopandas(gpd.GeoSeries(polygon))
-    gpdpoint = point_series.to_pandas()
-    gpdpolygon = polygon_series.to_pandas()
-    pd.testing.assert_series_equal(
-        gpdpolygon.contains(gpdpoint),
-        polygon_series.contains(point_series).to_pandas(),
+@pytest.mark.parametrize(
+    "point, polygon, expects",
+    [
+        [Point([-0.5, -0.5]), clockwiseSquare, False],
+        [Point([-0.5, 0.5]), clockwiseSquare, False],
+        [Point([0.5, 0.5]), clockwiseSquare, False],
+        [Point([0.5, -0.5]), clockwiseSquare, False],
+        # clockwise square, should be true
+        [Point([-0.5, 0.0]), clockwiseSquare, False],
+        [Point([0.0, 0.5]), clockwiseSquare, False],
+        [Point([0.5, 0.0]), clockwiseSquare, False],
+        [Point([0.0, -0.5]), clockwiseSquare, False],
+        # wound clockwise, should be false
+        [Point([0, 0]), clockwiseTriangle, False],
+        [Point([0.0, 1.0]), clockwiseTriangle, False],
+        [Point([1.0, 1.0]), clockwiseTriangle, False],
+        [Point([0.0, 0.5]), clockwiseTriangle, False],
+        [Point([0.5, 0.5]), clockwiseTriangle, False],
+        [Point([0.5, 1]), clockwiseTriangle, False],
+        # wound clockwise, should be true
+        [Point([0.25, 0.5]), clockwiseTriangle, True],
+        [Point([0.75, 0.9]), clockwiseTriangle, True],
+        # wound counter clockwise, should be false
+        [Point([0.0, 0.0]), Polygon([[0, 0], [1, 0], [1, 1], [0, 0]]), False],
+        [Point([1.0, 0.0]), Polygon([[0, 0], [1, 0], [1, 1], [0, 0]]), False],
+        [Point([1.0, 1.0]), Polygon([[0, 0], [1, 0], [1, 1], [0, 0]]), False],
+        [Point([0.5, 0.0]), Polygon([[0, 0], [1, 0], [1, 1], [0, 0]]), False],
+        [Point([0.5, 0.5]), Polygon([[0, 0], [1, 0], [1, 1], [0, 0]]), False],
+        # wound counter clockwise, should be true
+        [Point([0.5, 0.25]), Polygon([[0, 0], [1, 0], [1, 1], [0, 0]]), True],
+        [Point([0.9, 0.75]), Polygon([[0, 0], [1, 0], [1, 1], [0, 0]]), True],
+    ],
+)
+def test_point_in_polygon(point, polygon, expects):
+    gpdpoint = gpd.GeoSeries(point)
+    gpdpolygon = gpd.GeoSeries(polygon)
+    point = cuspatial.from_geopandas(gpdpoint)
+    polygon = cuspatial.from_geopandas(gpdpolygon)
+    result = polygon.contains(point)
+    assert gpdpolygon.contains(gpdpoint).values == result.values_host
+    assert result.values_host[0] == expects
+
+
+def test_two_points_one_polygon():
+    gpdpoint = gpd.GeoSeries([Point(0, 0), Point(0, 0)])
+    gpdpolygon = gpd.GeoSeries(Polygon([[0, 0], [1, 0], [1, 1], [0, 0]]))
+    point = cuspatial.from_geopandas(gpdpoint)
+    polygon = cuspatial.from_geopandas(gpdpolygon)
+    assert (
+        gpdpolygon.contains(gpdpoint).values
+        == polygon.contains(point).values_host
+    ).all()
+
+
+def test_one_point_two_polygons():
+    gpdpoint = gpd.GeoSeries([Point(0, 0)])
+    gpdpolygon = gpd.GeoSeries(
+        [
+            Polygon([[0, 0], [1, 0], [1, 1], [0, 0]]),
+            Polygon([[-2, -2], [-2, 2], [2, 2], [-2, -2]]),
+        ]
     )
+    point = cuspatial.from_geopandas(gpdpoint)
+    polygon = cuspatial.from_geopandas(gpdpolygon)
+    assert (
+        gpdpolygon.contains(gpdpoint).values
+        == polygon.contains(point).values_host
+    ).all()
+
+
+def test_ten_pair_points(point_generator, polygon_generator):
+    gpdpoints = gpd.GeoSeries([*point_generator(10)])
+    gpdpolygons = gpd.GeoSeries([*polygon_generator(10, 0)])
+    points = cuspatial.from_geopandas(gpdpoints)
+    polygons = cuspatial.from_geopandas(gpdpolygons)
+    assert (
+        gpdpolygons.contains(gpdpoints).values
+        == polygons.contains(points).values_host
+    ).all()
+
+
+def test_one_polygon_one_linestring(linestring_generator):
+    gpdlinestring = gpd.GeoSeries([*linestring_generator(1, 3)])
+    gpdpolygon = gpd.GeoSeries(Polygon([[0, 0], [0, 1], [1, 1], [0, 0]]))
+    linestring = cuspatial.from_geopandas(gpdlinestring)
+    polygons = cuspatial.from_geopandas(gpdpolygon)
+    assert (
+        gpdpolygon.contains(gpdlinestring).values
+        == polygons.contains(linestring).values_host
+    ).all()
+
+
+"""
+def test_onehundred_polygons_one_linestring(
+    linestring_generator,
+    polygon_generator
+):
+    gpdlinestring = gpd.GeoSeries([*linestring_generator(1, 3)])
+    gpdpolygons = gpd.GeoSeries([*polygon_generator(100, 0)])
+    linestring = cuspatial.from_geopandas(gpdlinestring)
+    polygons = cuspatial.from_geopandas(gpdpolygons)
+    assert (
+        gpdpolygons.contains(gpdlinestring).values
+        == polygons.contains(linestring).values_host
+    ).all()
+"""
